+++ src/js/tools/metroconfig.ts
@@ -1,6 +1,7 @@
-import type { MetroConfig } from 'metro';
+import type { MetroConfig, MixedOutput, Module, ReadOnlyGraph } from 'metro';
 
-import { createSentryMetroSerializer } from './sentryMetroSerializer';
+import { createSentryMetroSerializer, unstable_beforeAssetSerializationPlugin } from './sentryMetroSerializer';
+import type { DefaultConfigOptions } from './vendor/expo/expoconfig';
 
 export * from './sentryMetroSerializer';
 
@@ -10,13 +11,49 @@ export * from './sentryMetroSerializer';
  * Adds Debug ID to the output bundle and source maps.
  * Collapses Sentry frames from the stack trace view in LogBox.
  */
-export function withSentry(config: MetroConfig): MetroConfig {
+export function withSentryConfig(config: MetroConfig): MetroConfig {
   let newConfig = config;
 
   newConfig = withSentryDebugId(newConfig);
   newConfig = withSentryFramesCollapsed(newConfig);
 
   return newConfig;
+}
+
+/**
+ * This function returns Default Expo configuration with Sentry plugins.
+ */
+export function getSentryExpoConfig(projectRoot: string, options: DefaultConfigOptions = {}): MetroConfig {
+  const { getDefaultConfig } = loadExpoMetroConfigModule();
+  const config = getDefaultConfig(projectRoot, {
+    ...options,
+    unstable_beforeAssetSerializationPlugins: [
+      ...(options.unstable_beforeAssetSerializationPlugins || []),
+      unstable_beforeAssetSerializationPlugin,
+    ],
+  });
+
+  return withSentryFramesCollapsed(config);
+}
+
+function loadExpoMetroConfigModule(): {
+  getDefaultConfig: (
+    projectRoot: string,
+    options: {
+      unstable_beforeAssetSerializationPlugins?: ((serializationInput: {
+        graph: ReadOnlyGraph<MixedOutput>;
+        premodules: Module[];
+        debugId?: string;
+      }) => Module[])[];
+    },
+  ) => MetroConfig;
+} {
+  try {
+    // eslint-disable-next-line @typescript-eslint/no-var-requires
+    return require('expo/metro-config');
+  } catch (e) {
+    throw new Error('Unable to load `expo/metro-config`. Make sure you have Expo installed.');
+  }
 }
 
 type MetroCustomSerializer = Required<Required<MetroConfig>['serializer']>['customSerializer'] | undefined;
@@ -46,7 +83,9 @@ type MetroCustomizeFrameReturnValue =
 function withSentryFramesCollapsed(config: MetroConfig): MetroConfig {
   const originalCustomizeFrame = config.symbolicator?.customizeFrame;
   const collapseSentryInternalFrames = (frame: MetroFrame): boolean =>
-    typeof frame.file === 'string' && frame.file.includes('node_modules/@sentry/utils/cjs/instrument.js');
+    typeof frame.file === 'string' &&
+    (frame.file.includes('node_modules/@sentry/utils/cjs/instrument.js') ||
+      frame.file.includes('node_modules/@sentry/utils/cjs/logger.js'));
 
   const customizeFrame = (frame: MetroFrame): MetroCustomizeFrameReturnValue => {
     const originalOrSentryCustomizeFrame = (